Refactor: Split analyze_issue.py into modular package #276
Refactored analyze_issue.py from 1,352-line monolithic file into modular
package structure. All new files are under 700-line limit.

## Changes

### New Package Structure (src/vibe_check/tools/issue_analysis/)
- **models.py** (23 lines): IssueLabel dataclass for type safety
- **github_client.py** (98 lines): GitHubIssueClient wrapper for GitHub API
- **response_builder.py** (332 lines): AnalysisResponseBuilder for formatting
- **analyzer.py** (520 lines): EnhancedGitHubIssueAnalyzer with composition
- **api.py** (371 lines): Public API functions (analyze_issue, etc.)
- **__init__.py** (38 lines): Package initialization and exports

### Backward Compatibility
- **analyze_issue.py** (40 lines): Thin compatibility shim
- Re-exports all public API from issue_analysis package
- Maintains all existing import paths

### Architecture Changes
- Applied composition pattern over inheritance
- Separated concerns: GitHub client, response building, analysis logic
- Maintained backward compatibility with existing code
- Updated tests to work with new composition-based structure

### Test Updates
- Fixed test_github_issue_analyzer.py mocks for composition pattern
- Updated test_analyze_issue_workflow.py patch paths

## File Size Compliance
- analyze_issue.py: 1,352 → 40 lines (97% reduction)
- Largest new file: analyzer.py at 520 lines (26% under limit)
- All files comply with 700-line limit

Resolves #276
